% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_methods.R
\name{show_methods}
\alias{show_methods}
\title{Available ROC estimation methods}
\usage{
show_methods()
}
\value{
A \code{\link[data.table]{data.table}} containing the number and a short description of each implemented method.
}
\description{
Show the implemented ROC estimation methods.
}
