% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{variance}
\alias{variance}
\alias{variance.default}
\title{Sample Variance}
\usage{
variance(x, ...)

\method{variance}{default}(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{Further parameters.}
}
\description{
Sample Variance
}
\seealso{
\code{\link{variance.rkt_ecdf}}, \code{\link{var}}
}
